introduce new variant to `Special` enum
to keep all ranges of behavior possible a special variant to the
`Special` enum was introduced (ExcludeBosAndEos). It allows decoding of tokens but excludes
`bos` and `eos` tokens from the stream. This can be used to keep the old
behavior of llama-cpp-rs for decoding streams while allowing the
expected behavior that all tokens will be decoded by default.

See https://github.com/utilityai/llama-cpp-rs/pull/856 for the
discussion about this.